requirements.txt
robotframework==6.1.1
robotframework-seleniumlibrary==6.1.0
selenium==4.20.0
webdriver-manager==4.0.1
robotframework-requests==0.9.7
